Move tile cache calculation to ImageRegionStoreFactory
