+++ qupath-gui-fx/src/main/java/qupath/lib/gui/QuPathGUI.java
@@ -234,7 +234,7 @@ public class QuPathGUI {
 	/**
 	 * Default region store used by viewers for tile caching and repainting
 	 */
-	private DefaultImageRegionStore imageRegionStore = ImageRegionStoreFactory.createImageRegionStore(QuPathGUI.getTileCacheSizeBytes());
+	private DefaultImageRegionStore imageRegionStore = ImageRegionStoreFactory.createImageRegionStore();
 	
 	private MultiviewManager viewerManager = new MultiviewManager(this);
 	
@@ -3116,34 +3116,6 @@ public class QuPathGUI {
 			}
 		}
 	}
-	
-	
-	
-	/**
-	 * Calculate the appropriate tile cache size based upon the user preferences.
-	 * @return tile cache size in bytes
-	 */
-	private static long getTileCacheSizeBytes() {
-		// Try to compute a sensible value...
-		Runtime rt = Runtime.getRuntime();
-		long maxAvailable = rt.maxMemory(); // Max available memory
-		if (maxAvailable == Long.MAX_VALUE) {
-			logger.warn("No inherent maximum memory set - for caching purposes, will assume 64 GB");
-			maxAvailable = 64L * 1024L * 1024L * 1024L;
-		}
-		double percentage = PathPrefs.tileCachePercentageProperty().get();
-		if (percentage < 10) {
-			logger.warn("At least 10% of available memory needs to be used for tile caching (you requested {}%)", percentage);
-			percentage = 10;
-		} else if (percentage > 90) {
-			logger.warn("No more than 90% of available memory can be used for tile caching (you requested {}%)", percentage);
-			percentage = 90;			
-		}
-		long tileCacheSize = Math.round(maxAvailable * (percentage / 100.0));
-		logger.info(String.format("Setting tile cache size to %.2f MB (%.1f%% max memory)", tileCacheSize/(1024.*1024.), percentage));
-		return tileCacheSize;
-	}
-
 	
 	
 	/**
